+++ src/frontend/apps/desk/src/features/mail-domains/access-management/components/AccessesGrid.tsx
@@ -106,8 +106,6 @@ export const AccessesGrid = ({
 
   return (
     <Card
-      $padding={{ bottom: 'small' }}
-      $margin={{ all: 'big', top: 'none' }}
       $overflow="auto"
       $css={`
           & .c__pagination__goto {