+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to
 - ✨(api) add RELEASE version on config endpoint #459
 - ✨(backend) manage roles on domain admin view
 - ✨(frontend) show version number in footer #369
+- ✨(frontend) add tabs inside #466
 
 ### Changed